Merge model-controller connection to main

+++ src/main/resources/scripts/simulationMetrics.sql
@@ -3,9 +3,9 @@ USE simulation;
 DROP TABLE IF EXISTS country_metrics;
 CREATE TABLE country_metrics
 (
-    id                INT PRIMARY KEY AUTO_INCREMENT,
+    id                BIGINT PRIMARY KEY AUTO_INCREMENT,
     day               INT    NOT NULL,
-    country_id        INT    NOT NULL,
+    country_id        BIGINT NOT NULL,
     population        BIGINT NOT NULL,
     money             DOUBLE NOT NULL,
     average_happiness DOUBLE NOT NULL,
@@ -16,10 +16,10 @@ CREATE TABLE country_metrics
 DROP TABLE IF EXISTS resource_metrics;
 CREATE TABLE resource_metrics
 (
-    id          INT PRIMARY KEY AUTO_INCREMENT,
+    id          BIGINT PRIMARY KEY AUTO_INCREMENT,
     day         INT    NOT NULL,
-    country_id  INT    NOT NULL,
-    resource_id INT    NOT NULL,
+    country_id  BIGINT NOT NULL,
+    resource_id BIGINT NOT NULL,
     quantity    INT    NOT NULL,
     value       DOUBLE NOT NULL,
     FOREIGN KEY (country_id) REFERENCES country (id),
@@ -29,10 +29,10 @@ CREATE TABLE resource_metrics
 DROP TABLE IF EXISTS resource_node_metrics;
 CREATE TABLE resource_node_metrics
 (
-    id              INT PRIMARY KEY AUTO_INCREMENT,
+    id              BIGINT PRIMARY KEY AUTO_INCREMENT,
     day             INT    NOT NULL,
-    country_id      INT    NOT NULL,
-    resource_id     INT    NOT NULL,
+    country_id      BIGINT NOT NULL,
+    resource_id     BIGINT NOT NULL,
     production_cost DOUBLE NOT NULL,
     max_capacity    INT    NOT NULL,
     tier            INT    NOT NULL,
